fix: unexpected not defined errors when song changed
