+++ src/classes/audio-analysis.ts
@@ -1,8 +1,5 @@
 import { Expose } from 'class-transformer';
-
-export type Ran<T extends number> = number extends T ? number : _Range<T, []>;
-type _Range<T extends number, R extends unknown[]> = R['length'] extends T ? R[number] : _Range<T, [R['length'], ...R]>;
-
+import { Ran } from './type-definitions.js';
 export class AudioAnalysis {
   @Expose()
   sections: Array<Section>;
